code restructure: relocate bookmark files
