+++ src/components/bookmarks/BookmarkCard.tsx
@@ -7,8 +7,8 @@ import { Link } from "react-router-dom";
 
 import { stopBookmarksRedux } from "../../models/etaObjects";
 import { useAppDispatch, useAppSelector } from "../../store";
+import { removeStopBookmark } from "../../store/bookmarks/slice";
 import { fluentStyles } from "../../styles/fluent";
-import { removeStopBookmark } from "./stopBookmarkSlice";
 
 export function BookmarkCard(props: { id: number }) {
   const id = props.id;
