+++ src/components/eta/FavouriteEta.tsx
@@ -3,7 +3,6 @@ import { useEffect, useState } from "react";
 import { Trans, useTranslation } from "react-i18next";
 import { Link } from "react-router-dom";
 
-import Bookmark from "../../features/bookmarks/Bookmark";
 import {
   LineStopEta,
   stopBookmarkWithEta,
@@ -12,6 +11,7 @@ import {
 import { EtaPredictionXml } from "../../models/etaXml";
 import { store, useAppSelector } from "../../store";
 import { settingsSelectors } from "../../store/settings/slice";
+import Bookmark from "../bookmarks/Bookmark";
 import { FetchXMLWithCancelToken } from "../fetch/fetchUtils";
 import { multiStopParser, multiStopUnifier } from "../parser/multiStopParser";
 import RawDisplay from "../rawDisplay/RawDisplay";
