+++ src/index.tsx
@@ -3,13 +3,13 @@ import { createRoot } from "react-dom/client";
 import { Provider } from "react-redux";
 import { RouterProvider } from "react-router";
 
-import { saveStopBookmarkState } from "./features/bookmarks/localstorage";
 import "./i18n/i18n";
 import "./index.css";
 import reportWebVitals from "./reportWebVitals";
 import { router } from "./routes/MainRouter";
 import * as serviceWorkerRegistration from "./serviceWorkerRegistration";
 import { store } from "./store";
+import { saveStopBookmarkState } from "./store/bookmarks/state";
 import { saveSettingsState } from "./store/settings/state";
 import { FluentTheme } from "./styles/FluentTheme";
